requirements.txt
streamlit
transformers
torch
pandas
matplotlib
scikit-learn

# audio to text packages
mutagen
eyed3
groq
pyannote.audio